issues/17 refactor(rename): namespace `UnibusEvent`, class name `Unibus`
