--- Assets/Plugins/UnibusExample/Scripts/ObserverText.cs
@@ -1,19 +1,0 @@
-﻿using UnityEngine;
-using UnityEngine.UI;
-using System.Collections;
-using Unibus;
-
-public class ObserverText : MonoBehaviour
-{
-    public string Tag = "Count";
-    private Text Text;
-
-    void OnEnable()
-    {
-        this.Text = this.GetComponent<Text>();
-        this.BindUntilDisable(Tag, (int count) =>
-            {
-                this.Text.text = count.ToString();
-            });
-    }
-}
